Fix for new odoc
This just renames the module `Migrate_parsetree_versions` to be
`Versions`. It was already exposed as `Migrate_parsetree.Versions`,
and this is not changed by this PR. The problem was that in the
generated alias file (migrate_parsetree__.ml-gen), dune puts in
a @canonical tag assuming that the module will be exposed with
the same filename - so previously that was:

    @canonical Migrate_parsetree.Migrate_parsetree_versions

but since it's actually exposed as `Migrate_parsetree.Versions`
this didn't resolve and odoc couldn't canonicalise references leading
to broken links.

Just renaming the module so it matches dune's expectations fixes
the problem.

Signed-off-by: Jon Ludlam <jon@recoil.org>